refactor: 유저용 함수 분리
