+++ src/index.js
@@ -1,7 +1,8 @@
+import { getUserNumber, resetUserInput } from "./user.js";
+
 export default function BaseballGame() {
   let computerNumber = "";
   const hint = document.getElementById("result");
-  const userInput = document.getElementById("user-input");
   const submitButton = document.getElementById("submit");
 
   // 컴퓨터 난수 생성
@@ -19,67 +20,6 @@ export default function BaseballGame() {
     }
 
     computerNumber = number;
-  };
-
-  // 3자리 숫자인지 확인
-  const checkLengthThree = number => {
-    if (number.length !== 3) {
-      alert("세자리의 수를 입력해주세요.");
-      return true;
-    } else {
-      return false;
-    }
-  };
-
-  // 중복된 숫자 사용했는지 확인
-  const checkOverlap = number => {
-    if (
-      number[0] === number[1] ||
-      number[1] === number[2] ||
-      number[0] === number[2]
-    ) {
-      alert("중복없는 세자리의 수를 입력해주세요.");
-      return true;
-    } else {
-      return false;
-    }
-  };
-
-  // 사용자가 입력한 수가 유효한지 확인
-  const checkUserNumber = number => {
-    if (checkLengthThree(number)) {
-      return false;
-    } else if (checkOverlap(number)) {
-      return false;
-    } else if (checkValidNumber(number)) {
-      return false;
-    } else {
-      return true;
-    }
-  };
-
-  // 1~9의 숫자인지 확인
-  const checkValidNumber = number => {
-    let i = 0;
-    for (; i < 3; i++) {
-      if (number[i] < "1" || number[i] > "9") {
-        alert("1~9까지의 수를 입력해주세요.");
-        return true;
-      }
-    }
-
-    return false;
-  };
-
-  // 사용자로부터 입력받기
-  const getUserNumber = () => {
-    let userNumber = userInput.value;
-
-    if (checkUserNumber(userNumber)) {
-      return userNumber;
-    } else {
-      return "unvalid";
-    }
   };
 
   // 스트라이크 갯수 확인
@@ -162,8 +102,8 @@ export default function BaseballGame() {
   // 게임 재시작(난수 재생성, 힌트 및 입력 초기화)
   const resetGame = () => {
     getComputerNumber();
+    resetUserInput();
     hint.innerHTML = "";
-    userInput.value = "";
   };
 
   resetGame();
